Move dimension to separate module
